+++ STDOS/Drivers/ADS1246.h
@@ -9,11 +9,11 @@ class CADS1246
         CADS1246(Pin pincs, Pin pinsck, Pin pindin, Pin pindout, InputPort& pinrd, Pin pinreset);       
         byte ReadReg(byte RegAddr);
         void WriteReg(byte RegAddr, byte da);
-        float Read(void); //返回-1,表示转换未完成
+        int Read(void); //返回-1,表示转换未完成
         void Init(void);
         byte Check(void); //AD检查，正常返回0
     private:
-        float decodead(byte *da);
+        int decodead(byte *da);
 	
 		CSoftSpi *pspi; //SPI接口
         InputPort *ppinrd; //数据准备好接口
